+++ packages/core/src/query/index.ts
@@ -4,3 +4,4 @@ export {
 } from './FluxTableMetaData'
 export {default as FluxResultObserver} from './FluxResultObserver'
 export {default as FluxTableColumn, ColumnType} from './FluxTableColumn'
+export * from './flux'