+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 1.4.0 [unreleased]
 
+### Features
+
+1. [#186](https://github.com/influxdata/influxdb-client-js/pull/186): Allow to create parameterized queries using flux tagged template.
+
 ## 1.3.0 [2020-05-15]
 
 ### Features